# docs/openapi/api-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lurus API V2 - Multi-Tenant Gateway",
...
    "version": "2.0.0",
    "contact": {
      "name": "Lurus Support",
      "email": "support@lurus.cn"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://lurus.cn/terms"
    }
  },
  "servers": [
    {
      "url": "https://hub.lurus.cn",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v2/{tenant_slug}/data-policy/retention": {
      "get": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "Get Content Retention",
        "description": "Reads the relay-content retention setting of the caller's own tenant: the platform default, the tenant override and the effective mode (Tenant admin only).",
        "operationId": "getContentRetentionV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          }
        ],
        "responses": {
          "200": {
            "description": "Retention view",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRetentionView"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "Set Content Retention",
        "description": "Sets the tenant retention mode. A tenant may tighten past the platform default but never loosen below it (400 RETENTION_CANNOT_LOOSEN). Audited.",
        "operationId": "putContentRetentionV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRetentionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated retention view",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRetentionView"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST, INVALID_RETENTION or RETENTION_CANNOT_LOOSEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v2/{tenant_slug}/data-policy/tokens/{id}/retention": {
      "put": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "Set Token Content Retention",
        "description": "Sets the retention mode of one token of the caller's tenant; it may not be looser than the tenant's effective mode. A token of another tenant answers 404. Audited.",
        "operationId": "putTokenContentRetentionV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "$ref": "#/components/parameters/TokenId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRetentionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST, INVALID_RETENTION or RETENTION_CANNOT_LOOSEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v2/{tenant_slug}/data-policy/rules": {
      "get": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "List Tenant Content Rules",
        "description": "Lists the caller's tenant content rules plus the built-in pattern names and the per-scope limits.",
        "operationId": "listContentRulesV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          }
        ],
        "responses": {
          "200": {
            "description": "Rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRuleList"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "Create Tenant Content Rule",
        "description": "Creates a content rule for the caller's own tenant (the tenant is never read from the body). 409 RULE_LIMIT when the per-scope cap is reached. Audited.",
        "operationId": "createContentRuleV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created rule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRule"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST or INVALID_RULE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "RULE_LIMIT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/{tenant_slug}/data-policy/rules/{id}": {
      "put": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "Update Tenant Content Rule",
        "description": "Updates a rule of the caller's tenant; an id belonging to another tenant or to the platform scope answers 404 RULE_NOT_FOUND. Audited.",
        "operationId": "updateContentRuleV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "$ref": "#/components/parameters/ContentRuleId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated rule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRule"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST or INVALID_RULE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Tenant Admin"
        ],
        "summary": "Delete Tenant Content Rule",
        "description": "Deletes a rule of the caller's tenant; cross-tenant ids answer 404. Audited.",
        "operationId": "deleteContentRuleV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "$ref": "#/components/parameters/ContentRuleId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v2/admin/content-rules": {
      "get": {
        "tags": [
          "Platform Admin"
        ],
        "summary": "List Platform Content Rules",
        "description": "Lists platform-scope content rules (Platform Admin only).",
        "operationId": "listPlatformContentRulesV2",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRuleList"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Platform Admin"
        ],
        "summary": "Create Platform Content Rule",
        "description": "Creates a platform-scope rule (scope is fixed server-side). Audited.",
        "operationId": "createPlatformContentRuleV2",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created rule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRule"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST or INVALID_RULE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "RULE_LIMIT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/content-rules/{id}": {
      "put": {
        "tags": [
          "Platform Admin"
        ],
        "summary": "Update Platform Content Rule",
        "description": "Updates a platform-scope rule; a tenant rule id answers 404. Audited.",
        "operationId": "updatePlatformContentRuleV2",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ContentRuleId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated rule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentRule"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST or INVALID_RULE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Platform Admin"
        ],
        "summary": "Delete Platform Content Rule",
        "description": "Deletes a platform-scope rule. Audited.",
        "operationId": "deletePlatformContentRuleV2",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ContentRuleId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OIDCBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token issued by the deployment's configured OIDC provider. Which provider that is, is deploy-time configuration; no specific vendor is part of this contract."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie for platform admin"
      }
    },
    "parameters": {
      "TenantSlug": {
        "name": "tenant_slug",
        "in": "path",
        "required": true,
        "description": "Tenant identifier (slug). On routes behind TenantSlugGuard the reserved value `~` means \"the tenant this session authenticated into\"; the server resolves it from the session, and it is what the console sends. An explicit slug must equal that tenant (403 TENANT_MISMATCH otherwise). `GET /user/me` reports the real slug as `tenant_slug`.",
        "schema": {
          "type": "string",
          "pattern": "^([a-z0-9-]+|~)$",
          "example": "lurus"
        }
      },
...
      "TokenId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "ContentRuleId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid request"
          }
        }
      },
...
      "ContentRetentionView": {
        "type": "object",
        "properties": {
          "platform_default": {
            "type": "string",
            "description": "Platform-wide retention mode"
          },
          "tenant": {
            "type": "string",
            "description": "Tenant override; empty means inherit"
          },
          "effective": {
            "type": "string",
            "description": "Mode in force: the strictest of the layers"
          }
        }
      },
      "ContentRetentionRequest": {
        "type": "object",
        "properties": {
          "content_retention": {
            "type": "string",
            "enum": [
              "",
              "full",
              "metadata_only",
              "none"
            ],
            "description": "Empty string means inherit"
          }
        },
        "required": [
          "content_retention"
        ]
      },
...
      "ContentRuleRequest": {
        "type": "object",
        "properties": {
          "ordinal": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "role_scope": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "pattern_type": {
            "type": "string"
          },
          "builtin": {
            "type": "string"
          },
          "pattern": {
            "type": "string"
          },
          "replacement": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "ContentRuleList": {
        "type": "object",
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentRule"
            }
          },
          "builtins": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "max_rules": {
            "type": "integer"
          },
          "max_pattern_len": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}